% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_json.R
\name{format_json_vector}
\alias{format_json_vector}
\title{Convert \code{R} objects to/from JSON}
\usage{
format_json_vector(
  x,
  null = c("list", "null"),
  na = c("null", "string"),
  auto_unbox = TRUE,
  dataframe = c("rows", "columns", "values"),
  matrix = c("rowmajor", "columnmajor"),
  Date = c("ISO8601", "epoch"),
  POSIXt = c("string", "ISO8601", "epoch", "mongo"),
  factor = c("string", "integer"),
  complex = c("string", "list"),
  raw = c("base64", "hex", "mongo"),
  digits = 4,
  json_verbatim = TRUE,
  force = FALSE,
  ...
)
}
\description{
Convert \code{R} objects to/from JSON
}
\examples{
format_json_vector(letters)
format_json_vector(list(list(a = 1), list(a = 1, b = 2)))
}
