% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got_chars.R
\docType{data}
\name{got_chars_json}
\alias{got_chars_json}
\title{Game of Thrones characters}
\format{
A JSON array of objects. Each object describes a character and has
the following keys:
\itemize{
\item \code{url}
\item \code{id}
\item \code{name}
\item \code{alive}
\item \code{titles}
\item \code{aliases}
\item \code{allegiances}
}
}
\source{
\url{https://anapioficeandfire.com}
}
\usage{
got_chars_json
}
\description{
A JSON with some basic information about some of the Game of Thrones
characters.
}
\examples{
got_chars_json
}
\keyword{datasets}
