% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{jq_do}
\alias{jq_do}
\title{Execute jq command}
\usage{
jq_do(x, ..., json2 = TRUE, .na = json_na_error(), slurp = FALSE)
}
\description{
Execute jq command
}
\examples{
jq_do('{"a": 1}', ".a")
jq_do(c('{"a": 1}', '{"a": 2}'), ".a")
jq_do(NA)
jq_do(NA, .na = NA)
}
