% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_helpers.R
\name{jq_get_key}
\alias{jq_get_key}
\title{jq_get_key("abc")
jq_get_key(c("abc", "def"))
jq_get_key(list("abc", 2))}
\usage{
jq_get_key(key)
}
\description{
jq_get_key("abc")
jq_get_key(c("abc", "def"))
jq_get_key(list("abc", 2))
}
