% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_helpers.R
\name{jq_get_path}
\alias{jq_get_path}
\title{jq_get_path(".abc")
jq_get_path(".abc.def"))
jq_get_path(".abc\link{2}")}
\usage{
jq_get_path(path)
}
\description{
jq_get_path(".abc")
jq_get_path(".abc.def"))
jq_get_path(".abc\link{2}")
}
