% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json2}
\alias{json2}
\alias{new_json2}
\alias{as_json2}
\alias{is_json2}
\title{importFrom methods setOldClass
Construct a json2 object}
\usage{
json2(x = character())

new_json2(x = character())

as_json2(x)

is_json2(x)
}
\description{
importFrom methods setOldClass
Construct a json2 object

new_json2()
new_json2('{"abc": 1}')
new_json2(c('{"abc": 1}', '{"def": 2}', '\link{1, 2, 3}', NA))
}
\examples{
json2()
json2('{"abc": 1}')
}
