% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{json_agg_array}
\alias{json_agg_array}
\title{Aggregate into a JSON array}
\usage{
json_agg_array(x)
}
\arguments{
\item{x}{Vector to collapse into JSON array.}
}
\value{
A json2 object.
}
\description{
Aggregate into a JSON array
}
\examples{
json_agg_array(1:3)
json_agg_array(json2(c('{"a": 1}', '{"b": 2}')))
}
