% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utils.R
\name{json_array_length}
\alias{json_array_length}
\title{Get array length of JSON arrays}
\usage{
json_array_length(x, path = NULL, wrap_scalars = FALSE)
}
\arguments{
\item{x}{Vector with JSON.}

\item{path}{Path}

\item{wrap_scalars}{Consider scalars as length one array?}
}
\value{
An integer vector of array lengths
}
\description{
Get array length of JSON arrays
}
\examples{
json_array_length(c(NA, "[1, 2, 3]", "[1, 2]"))
json_array_length(1, wrap_scalars = TRUE)
}
