% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_object.R
\name{json_build_object}
\alias{json_build_object}
\title{Build a JSON object}
\usage{
json_build_object(x, ...)
}
\description{
Build a JSON object
}
\examples{
json_build_object(
  '{"foo": 5, "bar": {"foo": [1, 2, 3], "bar": 3}}',
  f = .foo, bf1 = `.bar.foo | .[1]`, bb = `.bar.bar`
)
}
