% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{json_delete_path}
\alias{json_delete_path}
\title{Delete key}
\usage{
json_delete_path(x, path)
}
\description{
Delete key
}
\examples{
json_delete_path(x, path = "abc")
json_delete_path(x, path = c("abc", "def"))
json_delete_path(x, path = c("not_there", "def"))
}
