% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{json_flatten_query}
\alias{json_flatten_query}
\title{Flatten an array of JSON-objects or JSON-arrays}
\usage{
json_flatten_query(x)
}
\arguments{
\item{x}{A JSON vector.}
}
\description{
Flatten an array of JSON-objects or JSON-arrays
}
