% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{json_flatten_value}
\alias{json_flatten_value}
\title{Flatten an array of values}
\usage{
json_flatten_value(x, ptype = NULL, wrap_scalars = FALSE)
}
\arguments{
\item{x}{A JSON vector.}

\item{ptype}{Output type. If \code{NULL}, the default, the output type is
determined by computing the common type across all elements of \code{...}.}

\item{wrap_scalars}{Should scalars be wrapped?}
}
\description{
Flatten an array of values
}
\examples{
json_flatten_value(c('[1,2,3]', '[4]'))
}
