% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_extractors.R
\name{json_get}
\alias{json_get}
\title{Get value from path}
\usage{
json_get(x, path, json2 = TRUE)
}
\description{
Postgres:
\itemize{
\item key: '{"a": {"b":"foo"}}'::json->'a'
\item path: '{"a": {"b":{"c": "foo"}}}'::json#>'{a,b}'
jq:
\item key: .a
.\link{"a"}
\item path: getpath(\link{"a", "b"})
.a.b
}
}
