% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{json_get_query}
\alias{json_get_query}
\title{Extract an object/array from JSON}
\usage{
json_get_query(x, path, wrap_scalars = FALSE, default = NULL, na = NA)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{Path to element to extract.}

\item{wrap_scalars}{Should scalars be wrapped?}

\item{default}{Default value if path doesn't exist or element at path is
empty.}

\item{na}{Default value if element of \code{x} is \code{NA}.}
}
\value{
A \code{json2} vector.
}
\description{
Extract a JSON object or array at a given path. To extract a JSON value
use \code{\link[=json_get_value]{json_get_value()}}.
}
