% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{json_get_value}
\alias{json_get_value}
\title{Extract a value from JSON}
\usage{
json_get_value(
  x,
  path,
  ptype = NULL,
  default = NULL,
  na = NA,
  bigint_as_char = TRUE
)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{Path to element to extract.}

\item{ptype}{Output type. If \code{NULL}, the default, the output type is
determined by computing the common type across all elements of \code{...}.}

\item{default}{Default value if path doesn't exist or element at path is
empty.}

\item{na}{Default value if element of \code{x} is \code{NA}.}

\item{bigint_as_char}{Convert big integers to character?}
}
\value{
A vector with class given by \code{.ptype} and length equal to \code{x}.
}
\description{
Extract a scalar JSON value at a given path. To extract a JSON object
or array use \code{\link[=json_get_query]{json_get_query()}}.
}
\examples{
x1 <- '{"a": 1, "b": 2}'
json_get_value(x1, "$.a")
json_get_value(c(NA_character_, x1), "$.a")
json_get_value(c(NA_character_, x1), "$.a", na = 3)
json_get_value(x1, "$.c", default = "not there")
try(json_get_value(x1, "$.c"))

json_get_query('{"a": {"b": 1}}', "$.a")
}
