% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{json_has_paths}
\alias{json_has_paths}
\alias{json_has_path}
\alias{json_has_all_paths}
\alias{json_has_any_paths}
\alias{json_has_path1}
\alias{json_has_paths1}
\alias{json_has_all_paths1}
\alias{json_has_any_paths1}
\title{Does path exists?}
\usage{
json_has_paths(x, paths)

json_has_path(x, path)

json_has_all_paths(x, paths)

json_has_any_paths(x, paths)

json_has_path1(x, path)

json_has_paths1(x, paths)

json_has_all_paths1(x, paths)

json_has_any_paths1(x, paths)
}
\description{
Does path exists?
}
\details{
Postgres
\itemize{
\item single string: '{"a":1, "b":2}'::jsonb ? 'b'
\item multiple strings (and): '{"a":1, "b":2, "c":3}'::jsonb ?| array\link{'b', 'c'}
\item multiple strings (or): '{"a":1, "b":2, "c":3}'::jsonb ?| array\link{'b', 'c'}
}

jq
\itemize{
\item single string: jq '{"a":1, "b":2}' | 'has("b")'
\item multiple strings (and): jq '{"a":1, "b":2}' | 'has("b") and has("c")'
\item multiple strings (or): jq '{"a":1, "b":2}' | 'has("b") or has("c")'
}
}
\examples{
json_has_path(x, c("a", "y"))
json_has_path(x, list("a", "y"))

json_has_paths(x, list(c("a", "x"), c("a", "y")))
json_has_all_paths(x, list(c("a", "x"), c("a", "y")))
json_has_any_paths(x, list(c("a", "x"), c("a", "y")))


json_has_path1(x1, c("a", "y"))
json_has_path1(x1, list("a", "y"))

json_has_paths1(x1, list(c("a", "x"), c("a", "y")))
json_has_all_paths1(x1, list(c("a", "x"), c("a", "y")))
json_has_any_paths1(x1, list(c("a", "x"), c("a", "y")))
}
