% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{json_keys}
\alias{json_keys}
\title{Get keys of JSON object resp. array}
\usage{
json_keys(x)
}
\arguments{
\item{x}{A JSON vector}
}
\value{
A list of keys.
}
\description{
Get keys of JSON object resp. array
}
\examples{
json_keys(c(
  '{"a": 1, "b": 2}',
  '{"x": 1, "y": 2}',
  "[1, 2]"
))
}
