% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{json_modify}
\alias{json_modify}
\title{Update values}
\usage{
json_modify(x, ...)
}
\description{
Update values
}
\examples{
json_modify(
  x,
  abc = list(
    def = 20,
    xyz = 10
  ),
  new_key = "new value"
)
cannot update array element
}
