% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{json_path_exists}
\alias{json_path_exists}
\title{Does the path exist?}
\usage{
json_path_exists(x, path)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{A JSON path.}
}
\value{
A logical vector.
}
\description{
Does the path exist?
}
\examples{
json_path_exists(
  c(
    '{"a": 1}',
    '{"b": 2}',
    "[1, 2]",
    NA_character_
  ),
  "$.a"
)
}
