% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{json_paths}
\alias{json_paths}
\alias{json_paths1}
\title{Get paths}
\usage{
json_paths(x)

json_paths1(x)
}
\description{
Postgres: doesn't exist
jq: \code{paths(x)}
}
\examples{
x1 = c('{"a": {"x": 1}, "b": 2, "c": 3}')
x2 = c('{"a": {"x": 11, "y": 22}, "s": 12, "t": [1, 2, 3]}')
x <- c(x1, x2)
json_paths(x)

json_paths1(x1)
}
