% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{json_prettify}
\alias{json_prettify}
\alias{json_minify}
\title{Prettify/Minify a JSON vector}
\usage{
json_prettify(x, indent = 3)

json_minify(x)
}
\arguments{
\item{x}{A JSON vector.}

\item{indent}{number of spaces to indent.}
}
\value{
A \code{json2} vector.
}
\description{
A wrapper around \code{\link[jsonlite:prettify]{jsonlite::prettify()}} resp. \code{\link[jsonlite:prettify]{jsonlite::minify()}}.
}
\examples{
x <- c("[1,2,   3]", '{"a": 1, "b": 2}')
json_prettify(x)
json_minify(x)
}
