% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{json_set}
\alias{json_set}
\title{Update value}
\usage{
json_set(x, path, value, create = FALSE)
}
\description{
json: \link{{"f1": 1, "f2": null}, 2}
Postgres:
jsonb_set(x, '{0,f3}', '3')
jq:
key: x | (.\link{0}.f3|=3)
path: setpath(\link{0, "f3"}, 3)
}
\examples{
json_set(x, path = "abc", value = "new value")
json_set(x, path = c("abc", "def"), value = "new value")
json_set(x, path = "new_key", value = "new value")
json_set(x, path = "new key", value = "new value")
}
