% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{json_set_path}
\alias{json_set_path}
\title{json_set_path(x, path = "abc", value = "new value")
json_set_path(x, path = c("abc", "def"), value = "new value")
json_set_path(x, path = "new_key", value = "new value")
json_set_path(x, path = "new key", value = "new value")}
\usage{
json_set_path(x, path, value, create = FALSE)
}
\description{
json_set_path(x, path = "abc", value = "new value")
json_set_path(x, path = c("abc", "def"), value = "new value")
json_set_path(x, path = "new_key", value = "new value")
json_set_path(x, path = "new key", value = "new value")
}
