% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_type}
\alias{json_type}
\title{Query the JSON type}
\usage{
json_type(x)
}
\arguments{
\item{x}{A JSON vector}
}
\value{
A character vector of JSON types
}
\description{
Query the JSON type
}
\examples{
json_type(c(NA, '1', 'null', '[1,2]', '{"a": 1}'))
}
