% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_u}
\alias{json_u}
\title{Unbox a vector or data frame}
\usage{
json_u(x)
}
\arguments{
\item{x}{atomic vector of length 1, or data frame with 1 row.}
}
\value{
A singleton version of \code{x}.
}
\description{
Mark a vector of length one to not be wrapped in an array when formatted as
\code{JSON}. This is only a tiny wrapper around \code{jsonlite::unbox()} to avoid
conflict with \code{rlang::unbox()}.
}
\examples{
format_json(list(foo = 123))
format_json(list(foo = json_u(123)))
}
