% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utils.R
\name{json_wrap_scalars}
\alias{json_wrap_scalars}
\title{Wrap scalars in a JSON array}
\usage{
json_wrap_scalars(x)
}
\arguments{
\item{x}{A character or numeric vector.}
}
\value{
A \code{json2} vector similar to \code{x} with the JSON scalars wrapped as
JSON array.
}
\description{
Wrap scalars in a JSON array
}
\examples{
json_wrap_scalars(c('["a", "b"]', "c", "d"))
json_wrap_scalars(c(1, 2))
}
