% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{jsonr_has_paths}
\alias{jsonr_has_paths}
\alias{jsonr_has_path}
\alias{jsonr_has_all_paths}
\alias{jsonr_has_any_paths}
\alias{jsonr_has_path1}
\alias{jsonr_has_paths1}
\alias{jsonr_has_all_paths1}
\alias{jsonr_has_any_paths1}
\title{Does path exists?}
\usage{
jsonr_has_paths(x, paths)

jsonr_has_path(x, path)

jsonr_has_all_paths(x, paths)

jsonr_has_any_paths(x, paths)

jsonr_has_path1(x, path)

jsonr_has_paths1(x, paths)

jsonr_has_all_paths1(x, paths)

jsonr_has_any_paths1(x, paths)
}
\description{
Does path exists?
}
\details{
Postgres
\itemize{
\item single string: '{"a":1, "b":2}'::jsonb ? 'b'
\item multiple strings (and): '{"a":1, "b":2, "c":3}'::jsonb ?| array\link{'b', 'c'}
\item multiple strings (or): '{"a":1, "b":2, "c":3}'::jsonb ?| array\link{'b', 'c'}
}

jq
\itemize{
\item single string: jq '{"a":1, "b":2}' | 'has("b")'
\item multiple strings (and): jq '{"a":1, "b":2}' | 'has("b") and has("c")'
\item multiple strings (or): jq '{"a":1, "b":2}' | 'has("b") or has("c")'
}
}
\examples{
jsonr_has_path(x, ".a.y")

jsonr_has_paths(x, c(".a.x", ".a.y"))
jsonr_has_all_paths(x, c(".a.x", ".a.y"))
jsonr_has_any_paths(x, c(".a.x", ".a.y"))


jsonr_has_path1(x1, ".a.y")

jsonr_has_paths1(x1, c(".a.x", ".a.y"))
jsonr_has_all_paths1(x1, c(".a.x", ".a.y"))
jsonr_has_any_paths1(x1, c(".a.x", ".a.y"))
}
