% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{jsonr_keys}
\alias{jsonr_keys}
\title{Get keys}
\usage{
jsonr_keys(x, .na = json_na_error())
}
\description{
Postgres: \code{jsonb_object_keys(x)}
jq: \code{keys(x)}
}
\examples{
x1 <- c('{"a": {"x": 1}, "b": 2, "c": 3}')
x2 <- c('{"a": {"x": 11, "y": 22}, "s": 12, "t": [1, 2, 3]}')
x <- c(x1, x2)
jsonr_keys(x)
jsonr_keys(c(x, NA), .na = c("a", "b"))
jsonr_keys1(x1)
}
