% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{jsonr_paths}
\alias{jsonr_paths}
\alias{jsonr_paths1}
\title{Get paths}
\usage{
jsonr_paths(x)

jsonr_paths1(x)
}
\description{
Postgres: doesn't exist
jq: \code{paths(x)}
}
\examples{
x1 <- c('{"a": {"x": 1}, "b": 2, "c": 3}')
x2 <- c('{"a": {"x": 11, "y": 22}, "s": 12, "t": [1, 2, 3]}')
x <- c(x1, x2)
jsonr_paths(x)

jsonr_paths1(x1)
}
