% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_convert_value.R
\name{new_json_object}
\alias{new_json_object}
\alias{new_json_array}
\title{Prototype helpers}
\usage{
new_json_object()

new_json_array()
}
\value{
An empty \code{json2_object} resp. \code{json2_array}.
}
\description{
Create a JSON object/array prototype to pass to the \code{.ptype} argument in
some of the functions.
}
