% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{object}
\alias{object}
\title{Creating JSON objects}
\description{
Creating JSON objects
}
\section{Create a JSON object from \code{R} values}{


Given some \code{R} values you can easily create an object with them by putting
them into a named list and applying \code{format_json()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{id <- 1
x <- 1:3
y <- c("a", "b")

list(id = json_u(id), x = x, y = y) \%>\%
  format_json()

# or in some cases you might be interested in using `dataframe = "columns"`
tibble::tibble(
  x = 1:3,
  y = c("a", "b", "c")
) \%>\%
  format_json(dataframe = "columns")
}\if{html}{\out{</div>}}

To create multiple objects at once (basically a vectorised version) put the
values in a data frame and apply \code{format_json_rowwise()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{df <- tibble::tibble(
  id = 1,
  x = list(1:2, 3:4, 5),
  y = c("a", "b", "c")
)
format_json_rowwise(df)
}\if{html}{\out{</div>}}
}

