% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{parse_json}
\alias{parse_json}
\title{Convert JSON to an \R{} object}
\usage{
parse_json(
  x,
  simplifyVector = TRUE,
  simplifyDataFrame = FALSE,
  simplifyMatrix = FALSE,
  flatten = FALSE,
  bigint_as_char = TRUE,
  na = list(),
  ...
)
}
\arguments{
\item{x}{a scalar JSON character}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{simplifyDataFrame}{coerce JSON arrays containing only records (JSON objects) into a data frame}

\item{simplifyMatrix}{coerce JSON arrays containing vectors of equal mode and dimension into matrix or array}

\item{flatten}{automatically \code{\link{flatten}} nested data frames into a single non-nested data frame}

\item{bigint_as_char}{Parse big ints as character?}

\item{na}{Value to return \code{x} is \code{NA}}

\item{...}{arguments passed on to \code{\link[jsonlite:parse_json]{jsonlite::parse_json}}}
}
\description{
Convert JSON to an \R{} object
}
