% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{parse_json_vector}
\alias{parse_json_vector}
\title{Parse a vector of jsons into a list}
\usage{
parse_json_vector(
  x,
  simplifyVector = TRUE,
  simplifyDataFrame = FALSE,
  simplifyMatrix = FALSE,
  flatten = FALSE,
  bigint_as_char = TRUE,
  .na = json_na_error(),
  .null = NULL,
  ...
)
}
\description{
Parse a vector of jsons into a list
}
\examples{
parse_json_vector(x = c('"a"', '"b"'))
parse_json_vector(x = c('"a"', '["b", "c"]'))
parse_json_vector(x = c('"a"', NA), .na = 1)
}
