% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json}
\alias{read_json}
\title{Read JSON from disk or url}
\usage{
read_json(path, ...)
}
\arguments{
\item{path}{Path or connection to read from}

\item{...}{arguments passed on to \code{\link{parse_json}}}
}
\value{
A \R{} object. The type depends on the input but is usually a list or a
data frame.
}
\description{
Read JSON from disk or url
}
