% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_helpers.R
\name{vec_as_jq_path}
\alias{vec_as_jq_path}
\title{jq_path("abc")
jq_path(c("abc", "def"))
jq_path(list("abc", 2))}
\usage{
vec_as_jq_path(x)
}
\description{
jq_path("abc")
jq_path(c("abc", "def"))
jq_path(list("abc", 2))
}
