% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{vec_cast.character.json}
\alias{vec_cast.character.json}
\title{workaround so that \code{json_verbatim = TRUE} works for json2 objects}
\usage{
\method{vec_cast.character}{json}(x, to, ...)
}
\description{
workaround so that \code{json_verbatim = TRUE} works for json2 objects
}
