% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{vec_cast.json2.jqson}
\alias{vec_cast.json2.jqson}
\title{#' @method vec_cast.json2 list
#' @export
vec_cast.json2.list <- function(x, to, ..., x_arg = "x", to_arg = "to") {
vec_list_cast(x, to, x_arg = x_arg, to_arg = to_arg)
}}
\usage{
\method{vec_cast.json2}{jqson}(x, to, ...)
}
\description{
#' @method vec_cast.json2 vctrs_list_of
#' @export
vec_cast.json2.vctrs_list_of <- function(x, to, ..., x_arg = "x", to_arg = "to") {
vec_list_cast(x, to, x_arg = x_arg, to_arg = to_arg)
}
}
