% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{vec_proxy_equal.json2}
\alias{vec_proxy_equal.json2}
\title{Equality}
\usage{
\method{vec_proxy_equal}{json2}(x, ...)
}
\description{
Equality
}
\examples{
x <- new_json2('{"a": {"x": 1, "y": 2}, "b": 2, "c": 3}')
x2 <- new_json2(
  '{
    "b": 2,
    "c": 3,
    "a": {
     "y": 2,
     "x": 1
    }
  }'
)

# order and format of json doesn't matter
x == x2

# different values are recognized
y <- new_json2('{"a": {"x": 1}, "b": 2, "c": 4}')
x == y
}
\keyword{internal}
