% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vec_ptype2.json2}
\alias{vec_ptype2.json2}
\alias{vec_ptype2.json}
\alias{vec_ptype2.jqson}
\alias{vec_ptype2.pq_jsonb}
\alias{vec_ptype2.pq_json}
\title{Coercion}
\usage{
\method{vec_ptype2}{json2}(x, y, ..., x_arg = "", y_arg = "")

\method{vec_ptype2}{json}(x, y, ..., x_arg = "", y_arg = "")

\method{vec_ptype2}{jqson}(x, y, ..., x_arg = "", y_arg = "")

\method{vec_ptype2}{pq_jsonb}(x, y, ..., x_arg = "", y_arg = "")

vec_ptype2.pq_json(x, y, ..., x_arg = "", y_arg = "")
}
\arguments{
\item{x, y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}

\item{x_arg, y_arg}{Argument names for \code{x} and \code{y}. These are used
in error messages to inform the user about the locations of
incompatible types (see \code{\link[vctrs:stop_incompatible_type]{stop_incompatible_type()}}).}
}
\value{
A length-zero \R{} object.
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.
}
