% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_json.R
\name{write_json}
\alias{write_json}
\title{Write JSON}
\usage{
write_json(x, path, ...)
}
\arguments{
\item{x}{An object to write to disk.}

\item{path}{Path or connection to write to.}

\item{...}{arguments passed on to \code{\link{format_json}}}
}
\description{
Write JSON
}
\seealso{
\code{\link[=format_json]{format_json()}}, \code{\link[=format_json_rowwise]{format_json_rowwise()}}
}
