% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_json_rowwise.R
\name{format_json_rowwise}
\alias{format_json_rowwise}
\title{Convert each tbl row to JSON}
\usage{
format_json_rowwise(
  df,
  null = c("list", "null"),
  na = c("null", "string"),
  auto_unbox = TRUE,
  matrix = c("rowmajor", "columnmajor"),
  Date = c("ISO8601", "epoch"),
  POSIXt = c("string", "ISO8601", "epoch", "mongo"),
  factor = c("string", "integer"),
  complex = c("string", "list"),
  raw = c("base64", "hex", "mongo"),
  digits = 4,
  json_verbatim = TRUE,
  force = FALSE,
  rownames = FALSE,
  ...
)
}
\arguments{
\item{df}{A dataframe.}

\item{null}{how to encode NULL values within a list: must be one of 'null' or 'list'}

\item{na}{how to print NA values: must be one of 'null' or 'string'. Defaults are class specific}

\item{auto_unbox}{automatically \code{\link{unbox}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link{unbox}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{I()}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}

\item{matrix}{how to encode matrices and higher dimensional arrays: must be one of 'rowmajor' or 'columnmajor'.}

\item{Date}{how to encode Date objects: must be one of 'ISO8601' or 'epoch'}

\item{POSIXt}{how to encode POSIXt (datetime) objects: must be one of 'string', 'ISO8601', 'epoch' or 'mongo'}

\item{factor}{how to encode factor objects: must be one of 'string' or 'integer'}

\item{complex}{how to encode complex numbers: must be one of 'string' or 'list'}

\item{raw}{how to encode raw objects: must be one of 'base64', 'hex' or 'mongo'}

\item{digits}{max number of decimal digits to print for numeric values. Use \code{I()} to specify significant digits. Use \code{NA} for max precision.}

\item{json_verbatim}{Leave json as it is and do not encode it again?}

\item{force}{unclass/skip objects of classes with no defined JSON mapping}

\item{rownames}{For data.frames add a field \verb{_row} with the row name?}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\description{
Convert each tbl row to JSON
}
\examples{
format_json_rowwise(mtcars[1:3, ])

# use a dataframe column for nested objects
df <- data.frame(x = 1:2)
df$y <- data.frame(z = c("a", "b"))
format_json_rowwise(df)

if (require("dplyr", quietly = TRUE, warn.conflicts = FALSE)) {
  # often useful in mutate/transmute
  mtcars \%>\%
    transmute(json = format_json_rowwise(tibble(mpg, cyl, extra = tibble(disp, hp))))
}
}
