% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{jq_do}
\alias{jq_do}
\title{Known jqr bugs:
\itemize{
\item jqr::jq(c("1", "2")) # returns 12
\item jqr::jq(NA_character_) # errors
\item jqr::jq(c('{"a": 1}', '2'), ".a") # suppresses error and doesn't return anything for 2nd argument
}}
\usage{
jq_do(x, ..., json2 = TRUE)
}
\description{
Known jqr bugs:
\itemize{
\item jqr::jq(c("1", "2")) # returns 12
\item jqr::jq(NA_character_) # errors
\item jqr::jq(c('{"a": 1}', '2'), ".a") # suppresses error and doesn't return anything for 2nd argument
}
}
