% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json2}
\alias{json2}
\alias{new_json2}
\alias{as_json2}
\alias{is_json2}
\title{Construct a json2 object}
\usage{
json2(x = character())

new_json2(x = character())

as_json2(x)

is_json2(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
\code{new_json2()} is a fast, low-level constructor that takes a character vector.
\code{json2()} checks the input for validity.
\code{as_json2()} and \code{is_json2()} are simple forwarders to \code{vctrs::vec_cast()}
and \code{vctrs::vec_is()} respectively.
}
\examples{
json2()
json2('{"abc": 1}')

new_json2()
new_json2('{"abc": 1}')
new_json2(c('{"abc": 1}', '{"def": 2}', "[1, 2, 3]", NA))
}
