% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utils.R
\name{json_array_length}
\alias{json_array_length}
\title{Get array length of JSON arrays}
\usage{
json_array_length(x, path = NULL, wrap_scalars = FALSE)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{Path to element. This must be a valid
\href{https://goessner.net/articles/JsonPath/index.html#e2}{JSONpath} expression.
For example \verb{"$.a.b[0]} extracts the \code{1} in \verb{\{"a": \{"b": [1, 2]\}\}}.}

\item{wrap_scalars}{Consider scalars as length one array?}
}
\value{
An integer vector of array lengths
}
\description{
Get array length of JSON arrays
}
\examples{
json_array_length(c(NA, "[1, 2, 3]", "[1, 2]"))

# scalars produce an error unless `wrap_scalars` is `TRUE`
json_array_length(1, wrap_scalars = TRUE)
}
