% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utils.R
\name{json_array_types}
\alias{json_array_types}
\title{Get the type of array elements}
\usage{
json_array_types(x, path = NULL)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{Path to element. This must be a valid
\href{https://goessner.net/articles/JsonPath/index.html#e2}{JSONpath} expression.
For example \verb{"$.a.b[0]} extracts the \code{1} in \verb{\{"a": \{"b": [1, 2]\}\}}.}
}
\value{
A character vector of JSON types.
}
\description{
This is barely useful on its own but can be of some use in combination with
\code{\link[=json_array_length]{json_array_length()}}.
}
\examples{
json_array_types(c("[1, true]", '["a", [1]]'))
}
\seealso{
\code{\link[=json_type]{json_type()}}
}
