% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json_is_valid}
\alias{json_is_valid}
\alias{json_assert_valid}
\title{Assert vector is valid JSON.}
\usage{
json_is_valid(x)

json_assert_valid(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
\code{json_is_valid()} returns a vector of \code{TRUE} and \code{FALSE}.
\code{json_assert_valid()} either throws an error with information on the invalid
elements or returns \code{x} invisibly
}
\description{
Uses \code{\link[jsonlite:validate]{jsonlite::validate()}} under the hood.
}
\examples{
json_is_valid("[1, 2]")
json_is_valid("[1, 2")

json_assert_valid("[1, 2]")
\dontrun{
json_assert_valid("[1, 2")
}
}
