% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_build_object.R
\name{json_build_object}
\alias{json_build_object}
\title{Build a JSON object}
\usage{
json_build_object(x, ...)
}
\arguments{
\item{x}{Input JSON.}

\item{...}{Object specification.}
}
\value{
A json2
}
\description{
Build a JSON object
}
\examples{
json_build_object('{"a": 1, "b": 2}', y = .a)
}
