% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{json_delete}
\alias{json_delete}
\title{Delete key}
\usage{
json_delete(x, ...)
}
\arguments{
\item{x}{A JSON vector.}

\item{...}{Paths to delete.}
}
\description{
Delete key
}
\examples{
x <- c('{"a": 11, "b": {"x": 12}}', NA)

json_delete(x, "$.a")
json_delete(x, "$.a", "$.b")
json_delete(x, "$.b.x")

json_delete(x, "$.abc")
}
