% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{json_delete}
\alias{json_delete}
\title{Delete key}
\usage{
json_delete(x, ...)
}
\description{
Delete key
}
\examples{
x <- c('{"a": 11, "b": {"x": 12}}')

json_delete(x, ".a")
json_delete(x, .a, .b)

json_delete(x, ".abc")
}
