% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{json_extract}
\alias{json_extract}
\title{Extract an element from JSON}
\usage{
json_extract(
  x,
  path,
  ptype = NULL,
  default = NULL,
  na = NA,
  wrap_scalars = FALSE,
  bigint_as_char = TRUE
)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{Path to element to extract.}

\item{ptype}{Output type. If \code{NULL}, the default, the output type is
determined by computing the common type across all elements of \code{...}.}

\item{default}{Default value if path doesn't exist or element at path is
empty.}

\item{na}{Default value if element of \code{x} is \code{NA}.}

\item{wrap_scalars}{Should scalars be wrapped?}

\item{bigint_as_char}{Convert big integers to character?}
}
\value{
A vector with class given by \code{.ptype} and length equal to \code{x}.
}
\description{
Extract an element at a given path.
}
\examples{
x1 <- '{"a": 1, "b": 2}'
json_extract(x1, "$.a")

json_extract(c(NA_character_, x1), "$.a")
json_extract(c(NA_character_, x1), "$.a", na = 3)

try(json_extract(x1, "$.c"))
json_extract(x1, "$.c", default = "not there")

json_extract('{"a": {"b": 1}}', "$.a")
}
