% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{json_flatten}
\alias{json_flatten}
\title{Flatten an array of values}
\usage{
json_flatten(x, ptype = NULL, allow_scalars = FALSE, bigint_as_char = TRUE)
}
\arguments{
\item{x}{A JSON vector.}

\item{ptype}{Output type. If \code{NULL}, the default, the output type is
determined by computing the common type across all elements of \code{...}.}

\item{bigint_as_char}{Convert big integers to character?}
}
\description{
Flatten an array of values
}
\examples{
json_flatten(c(x = "[1, 2]", y = "[3]"))
}
