% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{json_keys}
\alias{json_keys}
\title{Get keys}
\usage{
json_keys(x)
}
\description{
Postgres: \code{jsonb_object_keys(x)}
jq: \code{keys(x)}
}
\examples{
x1 <- c('{"a": {"x": 1}, "b": 2, "c": 3}')
x2 <- c('{"a": {"x": 11, "y": 22}, "s": 12, "t": [1, 2, 3]}')
x <- c(x1, x2, NA)
json_keys(x)
json_keys1(x1)
}
