% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{json_merge}
\alias{json_merge}
\title{Merge two jsons}
\usage{
json_merge(x, y)
}
\description{
Merge two jsons
}
\examples{
# something like list_modify and list_merge?
json_merge('{"a": 1, "c": 3}', '{"a": 11, "b": 2}')
}
