% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_path.R
\name{json_path_exists}
\alias{json_path_exists}
\title{Does the path exist?}
\usage{
json_path_exists(x, path)
}
\arguments{
\item{x}{A JSON vector.}

\item{path}{Path to element. This must be a valid
\href{https://goessner.net/articles/JsonPath/index.html#e2}{JSONpath} expression.
For example \verb{"$.a.b[0]} extracts the \code{1} in \verb{\{"a": \{"b": [1, 2]\}\}}.}
}
\value{
A logical vector.
}
\description{
Does the path exist?
}
\examples{
json_path_exists(
  c(
    '{"a": 1}',
    '{"b": 2}',
    "[1, 2]",
    NA_character_
  ),
  "$.a"
)
}
