% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{json_unnest_longer}
\alias{json_unnest_longer}
\title{Unnest a JSON array column}
\usage{
json_unnest_longer(
  data,
  col,
  path = NULL,
  values_to = NULL,
  indices_to = NULL,
  keys_to = NULL,
  ptype = NULL,
  wrap_scalars = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{col}{JSON-column of arrays to extract components from.}

\item{path}{Path where to extract from.}

\item{values_to}{Name of column to store vector values. Defaults to \code{col}.}

\item{indices_to}{A string giving the name of column which will contain the
rownumber before unnesting.}

\item{keys_to}{A string giving the name of column which will contain the
object keys resp. the array indices.}

\item{ptype}{Output type. If \code{NULL}, the default, the output type is
determined by computing the common type across all elements of \code{...}.}
}
\description{
Unnest a JSON array column
}
\examples{
df <- tibble::tibble(json = discog_json)
df

item_df <- df \%>\%
  json_unnest_longer(
    "json",
    values_to = "item"
  )
item_df

item_df \%>\%
  json_unnest_longer(
    "item",
    path = c("$.basic_information.artists"),
    values_to = "artist",
    indices_to = "component_id"
  )
}
