% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsontools-package.R
\docType{package}
\name{jsontools-package}
\alias{jsontools}
\alias{jsontools-package}
\title{jsontools: Working with JSON Vectors}
\description{
A toolbox for working with JSON vectors similar to the functions 'Postgres' provides to work with JSON columns. It supports in parsing and formatting JSON, extracting data from JSON, and modifying JSON data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mgirlich/jsontools}
}

}
\author{
\strong{Maintainer}: Maximilian Girlich \email{maximilian.girlich@outlook.com}

}
\keyword{internal}
