% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{parse_json_vector}
\alias{parse_json_vector}
\title{Parse a vector of jsons into a list}
\usage{
parse_json_vector(
  x,
  flatten = FALSE,
  bigint_as_char = TRUE,
  .na = json_na_error(),
  .null = NULL,
  ...
)
}
\arguments{
\item{x}{a scalar JSON character}

\item{flatten}{automatically \code{\link{flatten}} nested data frames into a single non-nested data frame}

\item{bigint_as_char}{Parse big ints as character?}

\item{.na}{Value to return if \code{x} is \code{NA}. By default an error of class
\code{jsontools_error_na_json} is thrown.}

\item{.null}{Return the prototype of \code{.null} if \code{x} is \code{NULL}
or a zero length character}

\item{...}{arguments passed on to \code{\link[jsonlite:read_json]{jsonlite::parse_json}}}
}
\description{
Parse a vector of jsons into a list
}
\examples{
parse_json_vector(x = c('"a"', '"b"'))
parse_json_vector(x = c('"a"', '["b", "c"]'))
parse_json_vector(x = c('"a"', NA), .na = 1)
}
